[create] 基本的なページ遷移を作成
